alterando estrutura da tabela de products

+++ database/migrations/2022_03_19_161712_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('slug');
-            $table->double('price');
+            $table->float('price');
             $table->longText('description');
             $table->integer('stock');
             $table->string('photo');
